Refactored to work with RMI
