+++ exercise1/src/SES_main.java
@@ -1,0 +1,29 @@
+import java.rmi.Naming;
+
+public class SES_main {
+    public SES_main(){}
+
+    public static void main(String[] args){
+        int numProcesses = 3;
+        try{
+            int[][] destinations = {{1,2}, {}, {1}};
+            String[][] messages = {{"1", "2"}, {}, {"3"}};
+            int[][] delays = {{5000,0}, {}, {500}};
+
+            for (int i = 0; i < numProcesses; i++) {
+                SchiperEggliSandozRMI process = (SchiperEggliSandozRMI) Naming.lookup("rmi://localhost/SchiperEggliSandoz-" + i);
+                for (int j = 0; j < destinations[i].length; j++) {
+                    if(i==2){
+                        Thread.sleep(1000);
+                    }
+                    process.send(messages[i][j], destinations[i][j], delays[i][j]);
+                }
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+}
+
+
+
